fix(failover): narrow service-unavailable pattern to prevent false rate-limit classification
A bare "service unavailable" response from a proxy or CDN was being
classified as provider-overload (rate_limit), causing false "API rate
limit reached" warnings. This narrows the overloaded pattern to require
co-occurrence with explicit overload/capacity/high-demand indicators,
and reorders classifyFailoverReason() to check timeout patterns before
overloaded patterns so that transient network errors are not misclassified.

Fixes #32828
